mv random functions to mathrand.h
