+++ ENIGMAsystem/SHELL/Universal_System/Extensions/DataStructures/data_structures.cpp
@@ -26,7 +26,7 @@
  **                                                                              **
  \********************************************************************************/
 
-#include "Universal_System/scalar.h"
+#include "Universal_System/mathrand.h"
 
 #include <float.h>
 #include <algorithm>
@@ -54,13 +54,9 @@ template<typename t> bool tequal(t v1, t v2) { return v1 == v2; }
 template<> bool tequal(float v1, float v2)   { return fequal(v1, v2); }
 template<> bool tequal(double v1, double v2) { return fequal(v1, v2); }
 
-namespace enigma_user {
-  // why the fuck can't we just include mathnc.h here?
-  // for connecting ds_list_shuffle and ds_grid_shuffle
-  // to randomize as specified by GM docs
-  extern ma_scalar random(ma_scalar n);
+namespace enigma {
   static inline int random_integer(int x) {
-    return int(random(x)); // +1 breaks shuffle WTF??!?
+    return int(enigma_user::random(x));
   }
 }
 
@@ -478,7 +474,7 @@ class grid
     }
     void shuffle()
     {
-        random_shuffle(grid_array, grid_array + (xgrid*ygrid - 1), enigma_user::random_integer);
+        random_shuffle(grid_array, grid_array + (xgrid*ygrid - 1), enigma::random_integer);
     }
 };
 
@@ -1289,7 +1285,7 @@ void ds_list_sort(const unsigned int id, const bool ascend)
 void ds_list_shuffle(const unsigned int id)
 {
   //shuffles the values in the list into a random order
-  random_shuffle(ds_lists[id].begin(), ds_lists[id].end(), enigma_user::random_integer);
+  random_shuffle(ds_lists[id].begin(), ds_lists[id].end(), enigma::random_integer);
 }
 
 bool ds_list_exists(const unsigned int id)
